rebalance refactors and cleaning

+++ src/constants.rs
@@ -1,3 +1,3 @@
-pub const MIN_TICK: i64 = -162_000_000;
-pub const MAX_TICK: i64 = 180_000_000;
+pub const MIN_TICK: i32 = -162_000_000;
+pub const MAX_TICK: i32 = 180_000_000;
